+++ src/models/simple_gan.py
